Split code into separate classes
